Ensure calls to lwgeom->type use type of uint8_t
git-svn-id: http://svn.osgeo.org/postgis/trunk@8027 b70326c6-7e19-0410-871a-916f4a2858ee

+++ liblwgeom/g_util.c
@@ -131,7 +131,7 @@ uint8_t gflags(int hasz, int hasm, int geodetic)
 * Type == 0 in the case of the string "GEOMETRY" or "GEOGRAPHY".
 * Return LW_SUCCESS for success.
 */
-int geometry_type_from_string(const char *str, int *type, int *z, int *m)
+int geometry_type_from_string(const char *str, uint8_t *type, int *z, int *m)
 {
 	char *tmpstr;
 	int tmpstartpos, tmpendpos;
